requirements.txt
trafilatura>=2.0.0
langdetect>=1.0.9
requests>=2.31.0
beautifulsoup4>=4.12.0
google-genai>=1.0.0
